refactored some packages and fixing swagger
