+++ src/main/java/com/project/winter/CompositeKeys/UnhealthyToHealthyCompositeKey.java
@@ -1,4 +1,4 @@
-package CompositeKeys;
+package com.project.winter.CompositeKeys;
 
 import java.io.Serializable;
