+++ src/main/java/com/project/winter/DAO/BrandDAO.java
@@ -1,9 +1,8 @@
-package DAO;
+package com.project.winter.DAO;
 
-import DTO.Brand;
+import com.project.winter.DTO.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.util.List;
 
 @Repository
 public interface BrandDAO extends JpaRepository<Brand,Long>{
