+++ src/main/java/com/project/winter/DAO/FoodToBrandDAO.java
@@ -1,9 +1,8 @@
-package DAO;
+package com.project.winter.DAO;
 
-import DTO.FoodToBrand;
+import com.project.winter.DTO.FoodToBrand;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.util.List;
 
 @Repository
 public interface FoodToBrandDAO extends JpaRepository<FoodToBrand,Long>{
