+++ src/main/java/com/project/winter/DTO/Brand.java
@@ -1,4 +1,4 @@
-package DTO;
+package com.project.winter.DTO;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.hibernate.validator.constraints.NotBlank;
