+++ src/main/java/com/project/winter/DTO/Food.java
@@ -1,4 +1,4 @@
-package DTO;
+package com.project.winter.DTO;
 
 import org.hibernate.validator.constraints.NotBlank;
